DoFnのUnitTestのサンプルを追加した
refs #1
